requirements.txt
# Core dependencies
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Optional: For better date parsing (if you want to extract dates from emails)
python-dateutil>=2.8.2

